requirements.txt
requests==2.31.0
python-dotenv==1.0.0
elevenlabs==0.2.26 